+++ age/benches/throughput.rs
@@ -6,7 +6,7 @@ use std::io::{self, Write};
 const KB: usize = 1024;
 
 fn bench(c: &mut Criterion<CyclesPerByte>) {
-    let recipients = vec![SecretKey::generate().to_public()];
+    let recipient = SecretKey::generate().to_public();
     let mut group = c.benchmark_group("stream");
 
     for size in &[KB, 2 * KB, 4 * KB, 8 * KB, 16 * KB, 128 * KB] {
@@ -15,7 +15,7 @@ fn bench(c: &mut Criterion<CyclesPerByte>) {
         group.throughput(Throughput::Bytes(*size as u64));
 
         group.bench_function(BenchmarkId::new("encrypt", size), |b| {
-            let mut output = Encryptor::with_recipients(recipients.clone())
+            let mut output = Encryptor::with_recipients(vec![Box::new(recipient.clone())])
                 .wrap_output(io::sink())
                 .unwrap();